requirements.txt
facebook-business>=17.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
uvicorn>=0.30.0
anthropic>=0.40.0
flask>=3.0.0
requests>=2.31.0
Pillow>=10.0.0
gunicorn>=22.0.0
